Delete item from wishlist, avoid merge conflicts
