+++ app/src/components/user/WishList.jsx
@@ -1,24 +1,25 @@
-import React, { Component } from 'react'
+import React, { Component } from 'react';
+import Wish from './Wish';
 
 class WishList extends Component {
+
   render() {
     return (
       <div>
         <h2>I'm the Wish List!</h2>
-        {
-          this.props.wishList.map(wish => (
-            <ul key={wish}>
-              {/* wish(itemId) needs to target an Item name in state.item */}
-              <li>
-                {wish}
-                <button onClick={console.log(`removed ${wish} from wishlist`)}>X</button>
-              </li>
-            </ul>
-          ))
-        }
+        <ul>
+          {
+            this.props.wishList.map(wish => (
+              <Wish
+                key={wish}
+                wish={wish}
+              />
+            ))
+          }
+        </ul>
       </div>
     );
   }
 }
 
-export default WishList
+export default WishList;
